extract penalizer tests utility code into separate file
